CAS-1051 Leverage JPA2 pessimistic locking to avoid deadlocks caused by lock upgrade (escalation) cases like the following:
 1. Tx1 obtains SHARE lock on TGT row P prior to creation of child ST C1
 2. Tx2 obtains SHARE lock on TGT row P prior to creation of child ST C2
 3. Tx2 attempts to upgrade SHARE lock to EXCLUSIVE in order to update P
 4. Tx1 attempts to upgrade SHARE lock to EXCLUSIVE in order to update P

Since Tx2 and Tx1 depend on each other in a way that will never resolve
naturally, the database reports a deadlock.

By querying for TGTs with lock type PESSIMISTIC, an exclusive lock is
obtained for P _prior_ to step 1 and Tx2 blocks on Tx1 at step 2 in a way
that will resolve naturally.