screen.success.security message can be confusing
The current wording for screen.success.security causes some users to only read the first half of the instruction, getting as far as "For security please log out". The user then gets stuck signing in to be immediately told to sign out.

I think the qualifier that the instruction is for when they have finished should come first.